requirements.txt
# vcita Test Agent - Dependencies
# Browser automation
playwright==1.40.0

# Configuration
pyyaml==6.0.1

# Database
sqlite-utils==3.35.2

# CLI and output
rich==13.7.0

# Image processing (for screenshots)
Pillow==10.1.0

# HTML/DOM parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# Testing framework
pytest==7.4.3
pytest-playwright==0.4.4

# Async support
asyncio-throttle==1.0.2

# Date/time handling
python-dateutil==2.8.2

# File watching (for development)
watchdog==3.0.0
